clever_prog/ppfb/qr_decoder_service/requirements.txt
fastapi
uvicorn
pyzbar
pillow
python-multipart